requirements.txt
boto3==1.7.77
botocore==1.10.77
certifi==2018.8.13
chardet==3.0.4
diff-match-patch==20121119
Django==1.9.12
django-health-check==3.7.0
django-import-export==1.0.1
django-storages==1.6.6
djangorestframework==3.4.7
docutils==0.14
et-xmlfile==1.0.1
idna==2.7
jdcal==1.4
jmespath==0.9.3
odfpy==1.3.6
openpyxl==2.5.5
Pillow==5.2.0
python-dateutil==2.7.3
PyYAML==3.13
requests==2.19.1
s3transfer==0.1.13
six==1.11.0
tablib==0.12.1
unicodecsv==0.14.1
urllib3==1.23
xlrd==1.1.0
xlwt==1.3.0
